Email notifications now use absolute link
Fixes #51

+++ com_openhouse/administrator/components/com_openhouse/install/install.sql
@@ -36,7 +36,7 @@ CREATE TABLE IF NOT EXISTS `#__openhouse_offices` (
 CREATE TABLE IF NOT EXISTS `#__openhouse_houses` (
 	`openhouse_house_id` SERIAL,
 
-	`price` MEDIUMINT NOT NULL,
+	`price` INT(11) NOT NULL,
 	`bedrooms` VARCHAR(10) NOT NULL,
 	`bathrooms` VARCHAR(10) NOT NULL,
 	`address` VARCHAR(30) NOT NULL,
